+++ mytimer/params.py
@@ -4,6 +4,7 @@
 MY_TIMER_VERSION = "0.9"
 INPUT_ERROR_MESSAGE = "[Error] Wrong input"
 SOUND_ERROR_MESSAGE = "[Error] Unable to play sound"
+PROGRAMS_LIST_TEMPLATE = "{0}. `{1}` - {2}"
 INPUT_EXAMPLE = "Example: python -m mytimer --hour=1 --minute=1 --second=1"
 TIME_ELEMENTS = ["minute", "second", "hour"]
 MESSAGE_TEMPLATE = "Message: {0}"
@@ -111,6 +112,7 @@ PROGRAMS_MAP = {
         "alarm": 1,
         "alarm_repeat": 1,
         "face": 1,
+        "tone": 1,
         "message": "Japanese green tea (2 mins)"},
     "tea-bag": {
         "hour": 0,